Set up boilerplate for adding new engines and added new deps
