+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# Changelog
+
+## v0.2.0 (2021-12-10)
+
+- Added support for additional pdf to text extraction engines
+    - [pytesseract](https://pypi.org/project/pytesseract/)
+    - [textract](https://textract.readthedocs.io/en/stable/index.html)
+
+## v0.1.1 (2021-12-10)
+
+- v0.1.1 Python ^3.7 support by @SamEdwardes in https://github.com/SamEdwardes/spaCyPDFreader/pull/2
+
