+++ spacypdfreader/__init__.py
@@ -1,3 +1,3 @@
 from spacypdfreader.spacypdfreader import pdf_reader
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'
